requirements.txt
apache-airflow
requests
psycopg2-binary
python-dotenv
